+++ History.md
@@ -1,4 +1,22 @@
-## vNEXT
 
-## v0.1.0 (7/16/2014)
-* Initial package release to atmosphere.
+v0.2.0 / 2014-07-25
+==================
+
+  * wip change to old package system for now
+  * No sense in being able to clear region data.
+  * Simplified tests — moved failing test case to IDT.
+  * Added test + fix for not passing data into render.
+  * add clearRegionData function
+  * fix destroyRegions bug
+  * fix render method to properly set data
+  * add onComplete callback feature to beginRendering function
+  * fix destroy regions when setting new layout template
+  * wip - fix up rendering transactions
+  * change region property to _region so no conflict with region function
+  * Added some failing tests to demonstrate data inheritance.
+  * fix package.js file to work with new packaging system
+
+v0.1.0 / 2014-07-16
+==================
+
+  * Initial package release to atmosphere.